Added Duck Typing
I went into each of the underlying classes
and changes there method names into the
same for each one. This allow me to call the
same method in Trip with out worrying on
the class of the given preparer. This adds
abstraction to my code due to the fact
that in Trip it isn't as well defined what
is going on or to what.