Added to Bicycle
I went in and expanded on my orginal
code. With this I added in more factors
to the class,however, this adds in a lot
more flaws, or at least more than before.
This is due to the many resposibilties that
this class holds, and it can clearly be seen
that the class Bicycle knows far too much
or more than it should to be safe.