Added Bicycle Subclasses
I went in and added in subclasses
for two types of bicycles. The mountain
bike and the road bike. The problem with
this code, however, is it isn't very flexible
when it comes to the possibility of
change which should always be
considered because it's such a high
likely hood.