Updating Subclasses
There is some problems with the
translation, but the ideas remain.
When it comes to spliting jobs up
for similar classes and subclasses
you should always start at the most
concrete points and then allow the
super class to be more abstract as it
goes up the chain. This helps break
up the responsibilities more evenly
and effectivly in the long run.