Added Schedule Class
I added in a new class which is design to work with a start
and end date. This new class is then implemented into the
bicycle class. This is extremely useful because it allows for
the schedule class to work as it should, but also keeps
the bicycle class from understanding more than it should.