Added Spares Class
This class is useful in slightly improving the code, but a big
topic it helps to show is how before the bicycle class knew
way too much about parts. This helped to put a boundary
around the two to prevent too much to be known about
each other, and to keep the code functioning with
different bike types.